Fix Stamina damage, show rounded AR totals and breakdown

+++ EldenRingBlazor/Data/Equipment/Weapon.cs
@@ -54,7 +54,7 @@
 
         public string Infusable { get; set; }
 
-        public int StaminaDamage { get; set; }
+        public double StaminaDamage { get; set; }
 
         public int Effect1 { get; set; }
